proper build fix
